refactor: send styles to single file in airdrop hook app
