--- apps/cowswap-frontend/src/modules/hooksStore/dapps/AirdropHookApp/styled/ContentWrapper.tsx
@@ -1,14 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const ContentWrapper = styled.div`
-  flex-grow: 1;
-  justify-content: center;
-  align-items: center;
-  flex-flow: column wrap;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  text-align: center;
-`
