--- apps/cowswap-frontend/src/modules/hooksStore/dapps/AirdropHookApp/styled/Input.tsx
@@ -1,21 +1,0 @@
-import { UI } from '@cowprotocol/ui'
-
-import styled from 'styled-components/macro'
-
-export const Input = styled.input`
-  border: 1px solid var(${UI.COLOR_BORDER});
-  background: var(${UI.COLOR_PAPER_DARKER});
-  color: inherit;
-  width: 100%;
-  margin: 10px 0;
-  margin-top: 0;
-  padding: 10px;
-  border-radius: 12px;
-  outline: none;
-  font-size: 15px;
-  font-weight: bold;
-
-  &:focus {
-    border: 1px solid var(${UI.COLOR_PRIMARY});
-  }
-`
