--- apps/cowswap-frontend/src/modules/hooksStore/dapps/AirdropHookApp/styled/LabelContainer.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const LabelContainer = styled.div`
-  width: fit-content;
-  margin: 10px 0;
-  label {
-    font-weight: 600;
-    font-size: 10pt;
-  }
-`
