--- apps/cowswap-frontend/src/modules/hooksStore/dapps/AirdropHookApp/styled/Link.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const Link = styled.button`
-  border: none;
-  padding: 0;
-  text-decoration: underline;
-  display: text;
-  cursor: pointer;
-  background: none;
-  color: white;
-  margin: 10px 0;
-`
