--- apps/cowswap-frontend/src/modules/hooksStore/dapps/AirdropHookApp/styled/Row.tsx
@@ -1,25 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const Row = styled.div`
-  display: flex;
-  flex-flow: column;
-  width: 100%;
-
-  label {
-    margin: 10px;
-    flex-grow: 0;
-    width: 5em;
-  }
-
-  input,
-  textarea {
-    flex-grow: 1;
-  }
-
-  /* Chrome, Safari, Edge, Opera */
-  input::-webkit-outer-spin-button,
-  input::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-`
