--- apps/cowswap-frontend/src/modules/hooksStore/dapps/AirdropHookApp/styled/Wrapper.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const Wrapper = styled.div`
-  display: flex;
-  flex-flow: column wrap;
-  width: 100%;
-
-  padding-bottom: 1rem;
-
-  flex-grow: 1;
-`
